Implemented authenication and authorization.

+++ internal/controllers/controller.go
@@ -3,7 +3,6 @@ package controllers
 
 import (
     "net/http"
-    "strings"
 
     "github.com/bartmika/mulberry-server/internal/repositories"
 )
@@ -23,13 +22,13 @@ func NewBaseHandler(u *repositories.UserRepo, tsd *repositories.TimeSeriesDatumR
 func (h *BaseHandler) HandleRequests(w http.ResponseWriter, r *http.Request) {
     w.Header().Set("Content-Type", "application/json")
 
-    // Split path into slash-separated parts, for example, path "/foo/bar"
-    // gives p==["foo", "bar"] and path "/" gives p==[""]. Our API starts with
-    // "/api/v1", as a result we will start the array slice at "3".
-    p := strings.Split(r.URL.Path, "/")[3:]
+    // Get our URL paths which are slash-seperated.
+    ctx := r.Context()
+    p := ctx.Value("url_split").([]string)
     n := len(p)
 
-    // fmt.Println(p, n) // For debugging purposes only.
+    // Get our authorization information.
+    isAuthorized := ctx.Value("is_authorized").(bool)
 
     switch {
     case n == 1 && p[0] == "version" && r.Method == http.MethodGet:
@@ -39,15 +38,35 @@ func (h *BaseHandler) HandleRequests(w http.ResponseWriter, r *http.Request) {
     case n == 1 && p[0] == "register" && r.Method == http.MethodPost:
         h.postRegister(w, r)
     case n == 1 && p[0] == "time-series-data" && r.Method == http.MethodGet:
-        h.getTimeSeriesData(w, r)
+        if isAuthorized {
+            h.getTimeSeriesData(w, r)
+        } else {
+            http.Error(w, "Unauthorized", http.StatusUnauthorized)
+        }
     case n == 1 && p[0] == "time-series-data" && r.Method == http.MethodPost:
-        h.postTimeSeriesData(w, r)
+        if isAuthorized {
+            h.postTimeSeriesData(w, r)
+        } else {
+            http.Error(w, "Unauthorized", http.StatusUnauthorized)
+        }
     case n == 2 && p[0] == "time-series-datum" && r.Method == http.MethodGet:
-        h.getTimeSeriesDatum(w, r, p[1])
+        if isAuthorized {
+            h.getTimeSeriesDatum(w, r, p[1])
+        } else {
+            http.Error(w, "Unauthorized", http.StatusUnauthorized)
+        }
     case n == 2 && p[0] == "time-series-datum" && r.Method == http.MethodPut:
-        h.putTimeSeriesDatum(w, r, p[1])
+        if isAuthorized {
+            h.putTimeSeriesDatum(w, r, p[1])
+        } else {
+            http.Error(w, "Unauthorized", http.StatusUnauthorized)
+        }
     case n == 2 && p[0] == "time-series-datum" && r.Method == http.MethodDelete:
-        h.deleteTimeSeriesDatum(w, r, p[1])
+        if isAuthorized {
+            h.deleteTimeSeriesDatum(w, r, p[1])
+        } else {
+            http.Error(w, "Unauthorized", http.StatusUnauthorized)
+        }
     default:
         http.NotFound(w, r)
     }
